feat: make useGradient hook
