+++ web/src/app/ee/admin/teams/Main.tsx
@@ -12,23 +12,7 @@ import {
 import { ThreeDotsLoader } from "@/components/Loading";
 import { useDocumentSets } from "@/app/admin/documents/sets/hooks";
 import { useParams } from "next/navigation";
-
-import config from "../../../../../tailwind-themes/tailwind.config";
-const tailwindColors = config.theme.extend.colors;
-
-const generateGradient = (teamspaceName: string) => {
-  const colors = [
-    tailwindColors.brand[100],
-    tailwindColors.brand[200],
-    tailwindColors.brand[300],
-    tailwindColors.brand[400],
-    tailwindColors.brand[500],
-  ];
-  const index = teamspaceName.charCodeAt(0) % colors.length;
-  return `linear-gradient(135deg, ${colors[index]}, ${
-    colors[(index + 1) % colors.length]
-  })`;
-};
+import { useGradient } from "@/hooks/useGradient";
 
 export const Main = ({ assistants }: { assistants: Assistant[] }) => {
   const { teamspaceId } = useParams();
@@ -88,7 +72,7 @@ export const Main = ({ assistants }: { assistants: Assistant[] }) => {
 
   const teamspacesWithGradients = data.map((teamspace) => ({
     ...teamspace,
-    gradient: generateGradient(teamspace.name),
+    gradient: useGradient(teamspace.name),
   }));
 
   const handleCloseSidebar = () => {
@@ -115,7 +99,6 @@ export const Main = ({ assistants }: { assistants: Assistant[] }) => {
       <TeamspaceSidebar
         assistants={assistants}
         selectedTeamspace={selectedTeamspace}
-        generateGradient={generateGradient}
         onClose={handleCloseSidebar}
         isExpanded={isExpanded}
         ccPairs={ccPairs}
