+++ web/src/components/globalSidebar/TeamspaceModal.tsx
@@ -9,9 +9,7 @@ import { MinimalTeamspaceSnapshot } from "@/lib/types";
 import Link from "next/link";
 import Image from "next/image";
 import { buildImgUrl } from "@/app/chat/files/images/utils";
-
-import config from "../../../tailwind-themes/tailwind.config";
-const tailwindColors = config.theme.extend.colors;
+import { useGradient } from "@/hooks/useGradient";
 
 interface TeamspaceModalProps {
   teamspace?: MinimalTeamspaceSnapshot[] | undefined;
@@ -27,20 +25,6 @@ export const TeamspaceModal = ({
   const [isModalVisible, setIsModalVisible] = useState(false);
 
   if (!teamspace) return null;
-
-  const generateGradient = (teamspaceName: string) => {
-    const colors = [
-      tailwindColors.brand[100],
-      tailwindColors.brand[200],
-      tailwindColors.brand[300],
-      tailwindColors.brand[400],
-      tailwindColors.brand[500],
-    ];
-    const index = teamspaceName.charCodeAt(0) % colors.length;
-    return `linear-gradient(135deg, ${colors[index]}, ${
-      colors[(index + 1) % colors.length]
-    })`;
-  };
 
   return (
     <CustomModal
@@ -86,7 +70,7 @@ export const TeamspaceModal = ({
               </div>
             ) : (
               <div
-                style={{ background: generateGradient(team.name) }}
+                style={{ background: useGradient(team.name) }}
                 className="font-bold text-inverted w-10 h-10 shrink-0 rounded-md bg-brand-500 flex justify-center items-center uppercase"
               >
                 {team.name.charAt(0)}
